+++ src/main/java/org/iesbelen/exception/AlumnoNotFoundException.java
@@ -1,7 +1,7 @@
 package org.iesbelen.exception;
 
 public class AlumnoNotFoundException extends RuntimeException{
-    public AlumnoNotFoundException(Long id) {
+    public AlumnoNotFoundException(int id) {
         super("Not found Alumno with id: " + id);
     }
 }
