+++ src/main/java/org/iesbelen/exception/AudicionNotFoundException.java
@@ -1,7 +1,7 @@
 package org.iesbelen.exception;
 
 public class AudicionNotFoundException extends RuntimeException{
-    public AudicionNotFoundException(Long id) {
+    public AudicionNotFoundException(int id) {
         super("Not found Audicion with id: " + id);
     }
 }
