+++ src/main/java/org/iesbelen/exception/ProfesorNotFoundException.java
@@ -1,7 +1,7 @@
 package org.iesbelen.exception;
 
 public class ProfesorNotFoundException extends RuntimeException{
-    public ProfesorNotFoundException(Long id) {
+    public ProfesorNotFoundException(int id) {
         super("Not found Profesor with id: " + id);
     }
 }
