+++ src/main/java/org/iesbelen/repository/AudicionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AudicionRepository extends JpaRepository<Audicion, Long> {
+public interface AudicionRepository extends JpaRepository<Audicion, Integer> {
 
 }
