+++ src/main/java/org/iesbelen/repository/ProfesorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Integer> {
 
     Profesor findByEmail(String email);
 }
